remove bug where colory will not build on *nix systems due to difference in data type size.
Internally, in the colory_init() function, it relies on many raw C DATA TYPES and also
defines its own to match those of the windows api. The size of data types in C may change
according to platform and that was causing the error. On windows systems, sizeof(unsigned long) is
4 bytes where it is usually 8 bytes on *nix systems.

Colory used the cfg!() macro to determine the platform. However it is said in the documentation-

cfg!, unlike #[cfg], does not remove any code and only evaluates to true or false.
For example, all blocks in an if/else expression need to be valid when cfg! is used for the condition, regardless of what cfg! is evaluating.

That means that the code would still compile even if it won't do anything.
However that also means that rust would check for data types and fail with mismatched data types.

Now, it instead uses the [cfg] attribute so that the some of the code wouldn't be includes on *nix systems and colory will build properly.